# Backend/swagger_server/models/branch.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Branch(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, branch_name: str=None, branch_code: str=None, branch_address: str=None):  # noqa: E501
        """Branch - a model defined in Swagger

        :param branch_name: The branch_name of this Branch.  # noqa: E501
        :type branch_name: str
        :param branch_code: The branch_code of this Branch.  # noqa: E501
        :type branch_code: str
        :param branch_address: The branch_address of this Branch.  # noqa: E501
        :type branch_address: str
        """
        self.swagger_types = {
            'branch_name': str,
            'branch_code': str,
            'branch_address': str
        }

        self.attribute_map = {
            'branch_name': 'branchName',
            'branch_code': 'branchCode',
            'branch_address': 'branchAddress'
        }
        self._branch_name = branch_name
        self._branch_code = branch_code
        self._branch_address = branch_address

    @classmethod
    def from_dict(cls, dikt) -> 'Branch':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Branch of this Branch.  # noqa: E501
        :rtype: Branch
        """
        return util.deserialize_model(dikt, cls)

    @property
    def branch_name(self) -> str:
        """Gets the branch_name of this Branch.


        :return: The branch_name of this Branch.
        :rtype: str
        """
        return self._branch_name

    @branch_name.setter
    def branch_name(self, branch_name: str):
        """Sets the branch_name of this Branch.


        :param branch_name: The branch_name of this Branch.
        :type branch_name: str
        """

        self._branch_name = branch_name

    @property
    def branch_code(self) -> str:
        """Gets the branch_code of this Branch.


        :return: The branch_code of this Branch.
        :rtype: str
        """
        return self._branch_code

    @branch_code.setter
    def branch_code(self, branch_code: str):
        """Sets the branch_code of this Branch.


        :param branch_code: The branch_code of this Branch.
        :type branch_code: str
        """

        self._branch_code = branch_code

    @property
    def branch_address(self) -> str:
        """Gets the branch_address of this Branch.


        :return: The branch_address of this Branch.
        :rtype: str
        """
        return self._branch_address

    @branch_address.setter
    def branch_address(self, branch_address: str):
        """Sets the branch_address of this Branch.


        :param branch_address: The branch_address of this Branch.
        :type branch_address: str
        """

        self._branch_address = branch_address